finished tutorial
suggested upgrades
For the current move only, show “You are at move #…” instead of a button
Rewrite Board to use two loops to make the squares instead of hardcoding them.
Add a toggle button that lets you sort the moves in either ascending or descending order.
When someone wins, highlight the three squares that caused the win (and when no one wins, display a message about the result being a draw).
Display the location for each move in the format (col, row) in the move history list.